///Version 0.8 Update///
Start lighthouse now actually works, it even checks for duplicate files. 
Just need to add the option to choose the behavior.

When pressing start lighthouse, a new thread is now started, so no more 
freezing of main window when the script is running.

The Quit button now actually stops everything, before the script would 
just continue running.

///In the next Update///

Next update is going to add the options for Lighthouse and behavior of 
duplicate files. Maybe even a overhaul of the GUI

Co-Authored-By: GreenSky-Productions <philippfilzer1999@gmail.com>